GG ELO/requirements.txt
# --- Core dependencies ---
discord.py>=2.3.2

# --- Async database support (e.g., aiosqlite or your custom backend) ---
aiosqlite>=0.19.0

# --- Image generation / leaderboard visuals ---
Pillow>=10.2.0

# --- Optional: For timezone/date parsing or formatting ---
python-dateutil>=2.8.2

# --- Optional: If you add translations (for multi-language support) ---
Babel>=2.14.0

# --- Optional: For rich console logging/debug (if desired) ---
rich>=13.7.0
